requirements.txt
# Web framework
Flask==3.0.0
gunicorn==21.2.0
flask-cors==4.0.0

# Google Cloud
google-cloud-ndb==2.1.0
firebase-admin==6.4.0
google-auth==2.23.4
google-cloud-datastore==2.19.0
six==1.16.0

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
flake8==6.1.0

# Environment and configuration
python-dotenv==1.0.0 